Service and Controllers commit

+++ backend/src/main/java/com/yourteam/cricketfantasy/service/UserService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface UserService {
     User createUser(User user);
-    User getUserById(Integer userId);
+    User getUserById(Long userId);
     List<User> getAllUsers();
-    User updateUser(Integer userId, User user);
-    void deleteUser(Integer userId);
+    User updateUser(Long userId, User user);
+    void deleteUser(Long userId);
 }
